[Saldo Revamp] Stage bug fixes on deposit and hold actvity

+++ features/fintech/saldo_details/src/main/java/com/tokopedia/saldodetails/response/model/saldoholdinfo/response/SaldoHoldDepositHistory.kt
@@ -8,7 +8,7 @@ data class SaldoHoldDepositHistory(
 	val sellerData: ArrayList<SaldoHoldInfoItem?>? = null,
 
 	@SerializedName("total")
-	val total: Int? = null,
+	val total: Double? = null,
 
 	@SerializedName("code")
 	val code: Int? = null,
